Refactor segmented sieve to use {2, 3, 5, 7}-wheel
Generate all p < 10^9 in 7.1s.

This implementation incurs a significant performance cost as it
allocates memory for all numbers in a segment instead of just the odds,
as in the previous wheel. The wheel spokes must moreover be initialized
to true.

TODO fix this memory + initialization cost by compressing the segment to
contain only spokes.