Extract sieve iterator state into its own struct
This slows the sieve to p < 10^9 in 2.6s, p < 10^10 in 27.4s.

The performance hit is likely due to interfering with compiler
optimization of origin_primes, which we've kept in the Sieve struct.
Now we're constructing a slice of origin_primes each time we want to
sieve a segment.

Although this change has a minor impact on performance, I would consider it
worthwhile for the improvement in readability.